requirements.txt
requests
urllib3==1.25.3
Django==2.2.2
mysqlclient==1.4.2.post1
PyGithub==1.43.7
djangorestframework==3.10.0
celery==4.3.0
django-cors-headers==3.0.2
django-filter==2.1.0
redis==3.2.1
pandas==0.25.1